+++ Player2/VideoCaptureDevice.cs
@@ -512,7 +512,7 @@ namespace fr.ipmfrance.webcam
                     
                     mediaControl = (IMediaControl) graphObject;
                     mediaEvent = (IMediaEventEx) graphObject;
-                    IntPtr p1, p2;
+                    int p1, p2;
                     DsEvCode code;
                     mediaControl.Run( );
